requirements.txt
Flask==3.0.0
python-consul==1.1.0
PyYAML==6.0.1
requests==2.31.0
APScheduler==3.10.4
